+++ src/consts.ts
@@ -3,7 +3,120 @@ import ogImage from '@/assets/images/purple.png';
 
 export const NAV_FOOTER: NavItem[] = [];
 export const NAV_MAIN: NavItem[] = [];
+export const NAV_TEST: NavItem[] = [
+	{
+		children: null,
+		description: 'Admin access test page',
+		href: '/test/admin',
+		permissions: ['admin'],
+		title: 'Admin Test',
+	},
+	{
+		children: null,
+		description: 'Editor access test page',
+		href: '/test/editor',
+		permissions: ['editor'],
+		title: 'Editor Test',
+	},
+	{
+		children: null,
+		description: 'Viewer access test page',
+		href: '/test/viewer',
+		permissions: ['viewer'],
+		title: 'Viewer Test',
+	},
+	{
+		children: null,
+		description: 'Component access test page',
+		href: '/test/components',
+		permissions: ['admin', 'editor', 'viewer'],
+		title: 'Components Test',
+	},
+];
 export const NAV_USER: NavItem[] = [];
+
+export const MENUS = {
+	NAV_FOOTER: [],
+	NAV_MAIN: [
+		{
+			children: null,
+			description: 'Home',
+			href: '/',
+			permissions: [],
+			title: 'Home',
+		},
+	],
+	NAV_TEST: [
+		{
+			children: [
+				{
+					children: null,
+					description: 'Admin access test page',
+					href: '/test/admin',
+					permissions: ['admin'],
+					title: 'Admin Test',
+				},
+				{
+					children: null,
+					description: 'Editor access test page',
+					href: '/test/editor',
+					permissions: ['editor'],
+					title: 'Editor Test',
+				},
+				{
+					children: null,
+					description: 'Viewer access test page',
+					href: '/test/viewer',
+					permissions: ['viewer'],
+					title: 'Viewer Test',
+				},
+				{
+					children: null,
+					description: 'Component access test page',
+					href: '/test/components',
+					permissions: ['admin', 'editor', 'viewer'],
+					title: 'Components Test',
+				},
+			],
+			description: 'A list of RBAC Test Pages',
+			href: '/test',
+			permissions: [],
+			title: 'RBAC Test',
+		},
+	],
+	NAV_SIGN: [
+		{
+			children: null,
+			description: 'Sign in',
+			href: '/sign-in',
+			permissions: [],
+			title: 'Sign in',
+		},
+		{
+			children: null,
+			description: 'Sign up',
+			href: '/sign-up',
+			permissions: [],
+			title: 'Sign up',
+		},
+	],
+	NAV_USER: [
+		{
+			children: null,
+			description: 'User dashboard',
+			href: '/u',
+			permissions: ['admin', 'editor', 'viewer'],
+			title: 'Dashboard',
+		},
+		{
+			children: null,
+			description: 'Admin area',
+			href: '/a',
+			permissions: ['admin'],
+			title: 'Admin',
+		},
+	],
+};
 
 export const SITE_DESCRIPTION =
 	'Astro builds fast content sites, powerful web applications, dynamic server APIs, and everything in-between.';